requirements.txt
pandas
chainlit
python-dotenv
pandasai
langchain-community
openai
